change position to PosInfo in orderidquery

+++ src/main/java/com/alibaba/middleware/race/model/PosInfo.java
@@ -12,10 +12,10 @@ package com.alibaba.middleware.race.model;
 public class PosInfo {
     public static final String POS_SPLITOR = "_";
     public static final String POS_SPLITOR_REX = "_";
-    public int offset;
+    public long offset;
     public int length;
     
-    public PosInfo(int offset, int length) {
+    public PosInfo(long offset, int length) {
         this.offset = offset;
         this.length = length;
     }
@@ -27,6 +27,6 @@ public class PosInfo {
 
     public static PosInfo parseFromString(String posInfoString) {
         String[] split = posInfoString.split(POS_SPLITOR_REX);
-        return new PosInfo(Integer.parseInt(split[0]), Integer.parseInt(split[1]));
+        return new PosInfo(Long.parseLong(split[0]), Integer.parseInt(split[1]));
     }
 }
